requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
websockets==12.0

# Azure dependencies
openai==1.3.5
azure-identity==1.15.0
azure-cosmos==4.5.1
azure-search-documents==11.4.0

# Environment and configuration
python-dotenv==1.0.0

# Async support
aiofiles==23.2.1
httpx==0.25.1

# Logging and monitoring
python-json-logger==2.0.7

# Optional: For production
# applicationinsights==0.11.10
# azure-monitor-opentelemetry==1.1.0